Starting to clean up code. API getters are now done in a hook
