+++ src/Types/Types.ts
@@ -61,3 +61,13 @@ export type StartTimeWithTheater = {
     startTime: Date;
     theater: Theater;
 }
+
+export type TheaterProps = {
+    type: string;
+    id: number;
+    implementationStrategy: void | null;
+    name: string;
+    totalRows: number;
+    seatsPerRow: number;
+    seats: Seat[];
+  }
