--- src/components/BookTicket/MovieShowDetail.json
@@ -1,32 +1,0 @@
-{
-    "movieshow": {
-      "id": "3",
-      "price": "200",
-      "start_date_time": "2023-10-09 21:00:00",
-      "movie": {
-        "id": "12",
-        "title": "Finding Nemo",
-        "poster": "https://image.tmdb.org/t/p/w500//ggQ6o8X5984OCh3kZi2UIJQJY5y.jpg",
-        "runtime": 100,
-        "voteRating": 7.824
-      },
-      "theatre": {
-        "type": "Standard",
-        "id": "1",
-        "seats_per_row": "20",
-        "total_rows": "10",
-        "cineman_id": "1"
-      },
-      "bookings": [
-        {
-          "id": "1",
-          "email": "lala@lala.com",
-          "seat": {
-            "id": "2",
-            "priceweigth": "1"
-          }
-        }
-      ]
-    }
-  }
-
