+++ src/views/LandingpageView.tsx
@@ -1,12 +1,15 @@
-import MainSection from '@/layouts/LandingpageLayout/MainSection'
+//import MainSection from '@/layouts/LandingpageLayout/MainSection'
 import React from 'react'
 
 const LandingpageView = () => {
   return (
     <div>
-        <MainSection />
+        
     </div>
   )
 }
 
+/*<div>
+        <MainSection />
+    </div>*/
 export default LandingpageView;
